Python_GCP_API/pyproject.toml
[tool.poetry]
name = "python-gcp-api"
version = "0.1.0"
description = ""
authors = ["GIORGIO PULCINI <giorgio.pulcini@alice.it>"]
packages = [
    { include = "python_gcp_api", from = "src" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-storage = "^2.16.0"
google-cloud-bigquery = "^3.20.1"
google-api-core = "^2.18.0"
pandas = "^2.2.2"
db-dtypes = "^1.2.0"
logging = "^0.4.9.6"
google-cloud-pubsub = "^2.21.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
script = "python_gcp_api.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
